refactor(Appointment): Doing the pages server side to handle the admin role in the appointments module
