+++ frontend/src/app/admin/appointments/[id]/page.jsx
@@ -1,15 +1,12 @@
-"use client";
 import Layout from "@/app/components/base_layout/Layout";
-import { useParams } from "next/navigation";
-import AppointmentDetail from "../../../forms/AppointmentDetail/AppointmentDetail";
+import { isAdmin } from "@/app/utils/requestBuilder";
+import AppointmentDetailAdmin from "./client";
 
-export default function AppointmentDetailAdmin() {
-  const { id } = useParams();
-
+export default async function Page() {
+  const isAdm = await isAdmin();
   return (
-    <Layout headerTitle={`Detalle de cita`}>
-      <title>SG BarberShop - Detalle de Cita</title>
-      <AppointmentDetail appointmentId={id}></AppointmentDetail>
+    <Layout headerTitle={`Detalle de cita`} isAdmin={isAdm}>
+      <AppointmentDetailAdmin/>
     </Layout>
   );
 }
