+++ frontend/src/app/admin/appointments/[id]/update/page.jsx
@@ -1,35 +1,13 @@
-"use client";
-import React, { useEffect, useState } from "react";
-import { getAppointment } from "../../../../apiHandlers/adminAppointments";
-import AppointmentForm from "../../../../forms/AppointmentForm/AppointmentForm";
 import Layout from "@/app/components/base_layout/Layout";
+import UpdateAppointment from "./client";
+import { isAdmin } from "@/app/utils/requestBuilder";
 
-export default function UpdateAppointment({ params }) {
-  const { id } = React.use(params);
-  const appointment = useAppointment(id);
+export default async function Page({ params }) {
+  const isAdm = await isAdmin();
 
   return (
-    <Layout mainTitle={"Actualizar cita"} headerTitle={"Modificar datos"}>
-      <title>SG BarberShop - Actualizar Cita</title>
-      {appointment ? (
-        <AppointmentForm appointment={appointment}></AppointmentForm>
-      ) : (
-        <p style={{ textAlign: "center" }}>Cargando...</p>
-      )}
+    <Layout mainTitle={"Actualizar cita"} headerTitle={"Modificar datos"} isAdmin={isAdm}>
+      <UpdateAppointment params={params}/>
     </Layout>
   );
-}
-
-function useAppointment(appointmentId) {
-  const [appointment, setAppointment] = useState(null);
-
-  useEffect(() => {
-    const fetchAppointment = async () => {
-      const data = await getAppointment(appointmentId);
-      setAppointment(data);
-    };
-
-    fetchAppointment();
-  }, []);
-  return appointment;
 }
